@@ -2,25 +2,25 @@
   "name": "WebApplicationBasic",
   "version": "0.0.0",
   "dependencies": {
-    "@angular/animations": "^4.1.2",
-    "@angular/common": "^4.1.2",
-    "@angular/compiler": "^4.1.2",
-    "@angular/core": "^4.1.2",
-    "@angular/forms": "^4.1.2",
-    "@angular/http": "^4.1.2",
-    "@angular/platform-browser": "^4.1.2",
-    "@angular/platform-browser-dynamic": "^4.1.2",
-    "@angular/platform-server": "^4.1.2",
-    "@angular/router": "^4.1.2",
-    "@progress/kendo-angular-dateinputs": "^1.0.4",
-    "@progress/kendo-angular-dropdowns": "^1.0.11",
-    "@progress/kendo-angular-excel-export": "^1.0.2",
-    "@progress/kendo-angular-grid": "^1.2.1",
-    "@progress/kendo-angular-inputs": "^1.0.5",
-    "@progress/kendo-angular-intl": "^1.2.0",
-    "@progress/kendo-angular-l10n": "^1.0.1",
-    "@progress/kendo-data-query": "^1.0.4",
-    "@progress/kendo-drawing": "^1.1.0",
+    "@angular/animations": "^5.0.1",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "@progress/kendo-angular-dateinputs": "^1.4.0",
+    "@progress/kendo-angular-dropdowns": "^1.2.3",
+    "@progress/kendo-angular-excel-export": "^1.0.5",
+    "@progress/kendo-angular-grid": "^1.6.1",
+    "@progress/kendo-angular-inputs": "^1.3.2",
+    "@progress/kendo-angular-intl": "^1.3.0",
+    "@progress/kendo-angular-l10n": "^1.0.5",
+    "@progress/kendo-data-query": "^1.1.1",
+    "@progress/kendo-drawing": "^1.4.0",
     "@progress/kendo-theme-default": "^2.37.3",
     "@types/node": "7.0.18",
     "angular2-template-loader": "0.6.2",
@@ -42,14 +42,14 @@
     "preboot": "4.5.2",
     "raw-loader": "0.5.1",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.0",
+    "rxjs": "^5.5.2",
     "style-loader": "0.17.0",
     "to-string-loader": "1.1.5",
-    "typescript": "2.3.2",
+    "typescript": "2.4.2",
     "url-loader": "0.5.8",
     "webpack": "2.5.1",
     "webpack-hot-middleware": "2.18.0",
     "webpack-merge": "4.1.0",
-    "zone.js": "0.8.10"
+    "zone.js": "^0.8.14"
   }
 }
